Ho cominciato la versione definitiva del codice

+++ load_dataset.py
@@ -48,7 +48,7 @@ def get_images(files_path, img_size_w, img_size_h, mode, randomize = False):
             # Controllo che le immagini abbiano tutte la stessa dimensione, altrimenti le salto
             if str(image.size) == '(' + str(img_size_w) + ', ' + str(img_size_h) + ')':
                 label_arr.append([0, 1])    # applico l'etichetta [0,1] a tutte le immagini SELECTION
-                images_arr.append(np.array(image,dtype=int))
+                images_arr.append(np.array(image,dtype=float))
             
     # Carico il NEUTRAL dataset [1,0]
     files = [f for f in listdir(neutral_path) if isfile(join(neutral_path, f))] 
@@ -61,7 +61,7 @@ def get_images(files_path, img_size_w, img_size_h, mode, randomize = False):
             # Controllo che le immagini abbiano tutte la stessa dimensione, altrimenti le salto
             if str(image.size) == '(' + str(img_size_w) + ', ' + str(img_size_h) + ')':
                 label_arr.append([1,0])         # applico l'etichetta [0,1] a tutte le immagini NEUTRAL
-                images_arr.append(np.array(image, dtype=int))
+                images_arr.append(np.array(image, dtype=float))
         
     # trasformo le liste in array numpy
     images_arr = np.array(images_arr)
